refactor(s3): rename S3 related class name and change epoch from 32bit to 64bit
1. rename S3 related class name
2. change epoch from 32bit to 64bit

+++ metadata/src/main/java/org/apache/kafka/metadata/stream/RangeMetadata.java
@@ -23,7 +23,7 @@ import org.apache.kafka.server.common.ApiMessageAndVersion;
 
 public class RangeMetadata implements Comparable<RangeMetadata> {
     private Long streamId;
-    private Integer epoch;
+    private Long epoch;
     private Integer rangeIndex;
     private Long startOffset;
     private Optional<Long> endOffset;
@@ -33,7 +33,7 @@ public class RangeMetadata implements Comparable<RangeMetadata> {
         return this.rangeIndex.compareTo(o.rangeIndex);
     }
 
-    public Integer getEpoch() {
+    public Long getEpoch() {
         return epoch;
     }
